Fix "Top" link in the footer of the default theme
Due to the use of the <base> tag, anchors (e.g. "#top") were relative
to the page defined in the <base> tag (the home page), rather than the
current page.

This meant that clicking on the "Top" link on e.g. the contact page would
load the home page, then scroll to the top of that. The result was a
confusing first impression for anyone playing with a default install of
the site.

The fix was to add {{ url:current }} before the #top in the anchor.