Added events call when calling build_forms
When creating your own module using the streams_core, and you want to override the default CP view for forms, and create your own, the events for each field type where not getting called. With this, now all the JS and CSS is getting called, and the functionality works perfectly.