Fixing issue with AJAX sorting in streams.
Previously, streams sorting ajax functions were housed under the Admin_Controller controller, so that only admins could sort entries and fields.

I've solved this for the 2.1 branch by checking to see if the current user has permission to access the module (at a base level) before allowing a sort. This is as far as we can go in 2.1, and we'll have more robust controls in 2.2.

Note: this will require a streams module upgrade.

This fixes #1513 and #1565.