Fixes an issue getting a field by slug
When retrieving a stream field by its' slug there is a conflict with the
cache where it doesn't respect the namespace that a field is in and
instead returns the first instance it finds in cache. This simply
updates the cache key to use namespace:slug to avoid this conflict.