New Admin Menu System
This commit removes the admin menu logic from the theme.php file and moves it into the Admin core controller. The admin menu is a core logic that should be present and consistent regardless of the admin theme.

This also includes a new way for modules to affect the menu. Since the menu is just an array, it can be manipulated with a function in the module's details.php file, or through an event. The order of items in the menu can also be manipulated.