Update to Module Upload Functionality.
This update stops a module from installing when it is uploaded. The reason behind this is that the user is hitting the "Upload" button, and it is assuming that the user wants to actually install it as well. "Upload" should simply mean "upload" so users can trust the semantics of the buttons.

Also, because of this change, we can now allow uploading of existing modules to update them, so users without FTP can update modules. They can then update them via the back end controls to update.

Fixes #2026.