Refining row parser join functionality.
This allows field types to do a join and add || as array delimiters. As an example, created_by has been baked into the row parser to be a join instead of a separate query.

This is on the edge of a new feature, but I consider the lack of advanced join functionality to be a bug more than anything, so I am counting it as a bug fix.

Closes #2179