Fix: Default page type when adding a new page
Currently page creation page fails to load if no page type is
specified.

This is an issue because users can get redirected to the creation page
without such a page type being specified.

One occasion where this happens is when they try to edit a page that
does not exist.

The proposed fix simply selects the first row from the page_type table
- avoiding the assumption that no one will delete the default
page_type.