Fixing issue with duplicating pages
The issue was that the plain ->get() method was being used for getting
the page data, which has a lot of logic in it to get/format entry data
from streams. This edit makes sure that just the bare bones data is
being grabbed.

Note: This will NOT work for fields that use alternative binding
processes, like Multiple Relationships. This should be addressed.

Fixes #2567